downloader: don't type constant

+++ eth/downloader/queue.go
@@ -41,7 +41,7 @@ const (
 
 var (
 	blockCacheMaxItems     = 8192             // Maximum number of blocks to cache before throttling the download
-	blockCacheInitialItems = uint64(2048)     // Initial number of blocks to start fetching, before we know the sizes of the blocks
+	blockCacheInitialItems = 2048             // Initial number of blocks to start fetching, before we know the sizes of the blocks
 	blockCacheMemory       = 64 * 1024 * 1024 // Maximum amount of memory to use for block caching
 	blockCacheSizeWeight   = 0.1              // Multiplier to approximate the average block size based on past ones
 )
